% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{adf}
\alias{adf}
\title{Quick and dirty chainr}
\usage{
adf(x)
}
\description{
\code{adf}
}
\seealso{
Other startup functions: \code{\link{chain.df}},
  \code{\link{dev.opts_in}}, \code{\link{packs.hush}},
  \code{\link{parse.file_ext}}, \code{\link{ul.as}}
}

