% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{chain.df}
\alias{chain.df}
\title{Pass an object through a dataframe converting chain}
\usage{
chain.df(x, col_name = NULL, row_name = NULL)
}
\description{
\code{chain.df} returns an object as a Data Frame.
}
\details{
Most of the time when manipulating data in a 'pipe'
 converting an object or array into a data frame is
 simplest, but comes with varying drawbacks and I
 don't normally set opitions(stringsAsFactors=F)
 as a default. This allows passing a melted or converted object
 in a chain and setting the class on each column as well as
 renaming the defaults or including the row names as a column
}
\examples{

> letters \%>\% chain.df('alphabet')

> A data frame with 26 observations on the following 1 variables.
> \\describe{
> \\item{alphabet}{a character vector}\\emph{ex:a}
> }

}
\seealso{
Other startup functions: \code{\link{adf}},
  \code{\link{dev.opts_in}}, \code{\link{packs.hush}},
  \code{\link{parse.file_ext}}, \code{\link{ul.as}}
}

