% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{dev.opts_in}
\alias{dev.opts_in}
\title{Set the variables for the DESCRIPTION files when building a new package}
\usage{
dev.opts_in(title_line = NULL, desc = NULL)
}
\arguments{
\item{title_line}{string The Top line description for the package itself}

\item{desc}{The longer explanation for the package. This should include
details, a main purpose, relatives and why it was built.}
}
\description{
\code{dev.opts_in}
}
\details{
This should be one of the first two scripts loaded and run with each
package. This one should be universal across apps for standard purposes.
}
\seealso{
Other startup functions: \code{\link{adf}},
  \code{\link{chain.df}}, \code{\link{packs.hush}},
  \code{\link{parse.file_ext}}, \code{\link{ul.as}}
}

