% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{dirs.file_exists}
\alias{dirs.file_exists}
\title{Find if a file exists in the local library}
\usage{
dirs.file_exists(name_hint = NULL, file_type = NULL,
  return_file_only = FALSE)
}
\arguments{
\item{name_hint}{string a hint or word contained in the file name.}

\item{file_type}{string The extension of the file type}

\item{return_file_only}{logical Should the reference data frame be returned or
just the file names and paths.}
}
\value{
A data frame
\itemize{
  \item files character R/data_parse/chains.R
 \item ext character .R
}

Additionally or optionally a character vector of all returned files or
if no items found a string of "No Files"
}
\description{
\code{dirs.file_exists}
}
\details{
This will find and locate any file and allow for filtering by a name hint
and a file extension type
}
\examples{
> dirs.file_exists(name_hint = "nvd3",file_type = "rhtml",TRUE)

# See if the icon reference data is present
> dirs.file_exists(name_hint = "icon",file_type = "rds",TRUE)

# All Files containing 'parse'
> dirs.file_exists("parse",return_file_only = TRUE)

}

