% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{dirs.files_in_lib}
\alias{dirs.files_in_lib}
\title{Find the folders given a directory}
\usage{
dirs.files_in_lib(lib_loc = NULL, is_r = TRUE, file_ext = NULL,
  file_hint = NULL, function_name = FALSE)
}
\description{
Find the folders given a directory
}

