% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{dirs.show_folders}
\alias{dirs.show_folders}
\title{Series of functions for working with directories}
\usage{
dirs.show_folders(r_dir = TRUE, dir_in = NULL)
}
\description{
\code{dirs.show_folders}
}
\details{
Get the top level folders from a location on the windows file directory
}

