% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_utils.R
\name{dirs.vis_studio}
\alias{dirs.vis_studio}
\title{Shortcut to working with VS}
\usage{
dirs.vis_studio(type = c("website", "project"), show_names = FALSE,
  find_lib = NULL, find_folder = NULL, find_file = NULL,
  find_file_ext = NULL)
}
\description{
\code{dirs.vis_studio}
}

