% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{doc.make_lic}
\alias{doc.make_lic}
\title{Generate a LICENSE formatted to Rd specs}
\usage{
doc.make_lic(owner_name = NULL, owner_org = NULL, year = "current",
  output_to = c("file", "console"))
}
\arguments{
\item{owner_name}{string Default to NULL Owner of the License.}

\item{owner_org}{string Default to NULL If the owner attributes
work to an Organization.}

\item{year}{character Default to current year.}

\item{output_to}{call Default to file and console Where should the output
print?}
}
\value{
Output is either a LICENSE file object in the local or working
 directory or printed to the console
}
\description{
\code{doc.make_lic}
}
\examples{
doc.make_lic(owner_name = 'Carl Boneri',owner_org = "FAS Advisory",
output_to = 'file')
#> #'@param owner_name string Default to NULL
#> #'@param owner_org string Default to NULL
#> #'@param year character Default to current
#> #'@param output_to call Default to file and console

}

