% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{drop.findr}
\alias{drop.findr}
\title{A function for search and return in dropbox.}
\usage{
drop.findr(..., file_only = T, file_ext = NULL)
}
\arguments{
\item{file_ext}{}
}
\value{
A character vector with either the full or
relative path or the specific file with full path
attached in the name for easy sourcing.
}
\description{
\code{drop.findr}
}
\details{
This will allow to recursively search large libraries in
dropbox and return specific files or directories meeting
input criteria.
}
\examples{
Returns a list of all directories containing kw icon
> drop.findr('icon',file_only = F)
[1] "/rterial/icons"         "/www/icon"  "/www/icon/flat_icons"
 "/www/icon/google_icons" "/www/icon/brand_icons"  "/www/icon/file_icons"

Returns a specific and best match file

drop.findr('icon',file_only = T,file_ext = 'png')
[1] "/www/icon/brand_icons/R_file_icon.png"



}

