% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_validations.R
\name{is.true}
\alias{is.true}
\title{FUNCTIONS
\code{is.true}}
\usage{
is.true(x)
}
\description{
\code{is.false}
}
\details{
\code{is.empty_list}

\code{is.real_null}

\code{is.real_na}

\code{swap.na}

\code{swap.null}

\code{swap.is_false}

\code{swap.is_true}

\code{swap.empty_list}

\code{swap.is_invalid}

\code{list.drop}

\code{list.drop_null}

\code{list.drop_na}

\code{list.drop_false}

\code{list.drop_true}

\code{list.drop_zero}

Validate truth of a conditional statement

\code{is.true}
}
\seealso{
\code{\link{is.true}}, \code{\link{is.false}},
\code{\link{is.empty_list}}, \code{\link{is.real_null}},
\code{\link{is.real_na}}, \code{\link{swap.na}}, \code{\link{swap.null}},
\code{\link{swap.is_false}}, \code{\link{swap.is_true}},
\code{\link{swap.empty_list}}, \code{\link{swap.is_invalid}},
\code{\link{list.drop}}, \code{\link{list.drop_null}},
\code{\link{list.drop_na}}, \code{\link{list.drop_false}},
\code{\link{list.drop_true}}, \code{\link{list.drop_zero}}
}

