% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_validations.R, R/startup_utils.R
\name{list.drop}
\alias{}
\alias{list.drop}
\title{Remove null or NA slots from lists}
\usage{
list.drop(x, drop_type = c("na", "null"))

list.drop(x, drop_type = c("na", "null"))
}
\arguments{
\item{drop_type}{option Either NA or NULL}

\item{drop_type}{option Either NA or NULL}
}
\description{
\code{list.drop}

\code{list.drop}
}
\examples{


}
\seealso{
\code{\link{dep_list}},
\code{\link{file_list}},
\code{\link{list.drop}}, \code{\link{list.drop_na}},
\code{\link{list.drop_null}},
\code{\link{mtrl.dep_to_list}},
\code{\link{packs.get_list}}, \code{\link{vig_list}}

\code{\link{dep_list}},
\code{\link{file_list}},
\code{\link{list.drop}}, \code{\link{list.drop_na}},
\code{\link{list.drop_null}},
\code{\link{mtrl.dep_to_list}},
\code{\link{packs.get_list}}, \code{\link{vig_list}}

Other data parse: \code{\link{list.drop_zero}}

Other data parse: \code{\link{list.drop_zero}}
}

