% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_validations.R, R/startup_utils.R
\name{list.drop_na}
\alias{list.drop_na}
\title{Short hand assignments for the \emph{list.drop_} family.}
\usage{
list.drop_na(x)

list.drop_na(x)
}
\description{
\code{list.drop_na}

\code{list.drop_na}
}

