% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_validations.R, R/startup_utils.R
\name{list.drop_zero}
\alias{list.drop_zero}
\title{Remove empty list objects returned in the fashion of 'list()'}
\usage{
list.drop_zero(x)

list.drop_zero(x)
}
\arguments{
\item{x}{A list or data frame object.}

\item{x}{A list or data frame object.}
}
\value{
Given a data object, return all objects whose length is greater
than zero and is not of NULL or NA object orientation.

Given a data object, return all objects whose length is greater
than zero and is not of NULL or NA object orientation.
}
\description{
\code{\link{list.drop_zero}}

\code{list.drop_zero}
}
\examples{

aa <- as.list(mtcars)

toString(names(aa))
[1] "mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb"

aa$gear <- list()

aa <- list.drop_zero(aa)

bb <- list.drop_zero(aa)

names(aa)[!names(aa) \%in\% names(bb)]

[1] "gear"


aa <- as.list(mtcars)

toString(names(aa))
[1] "mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb"

aa$gear <- list()

aa <- list.drop_zero(aa)

bb <- list.drop_zero(aa)

names(aa)[!names(aa) \%in\% names(bb)]

[1] "gear"

}
\seealso{
\code{\link{dep_list}},
\code{\link{file_list}},
\code{\link{list.drop}}, \code{\link{list.drop_na}},
\code{\link{list.drop_null}},
\code{\link{mtrl.dep_to_list}},
\code{\link{packs.get_list}}, \code{\link{vig_list}}

\code{\link{dep_list}},
\code{\link{file_list}},
\code{\link{list.drop}}, \code{\link{list.drop_na}},
\code{\link{list.drop_null}},
\code{\link{mtrl.dep_to_list}},
\code{\link{packs.get_list}}, \code{\link{vig_list}}

Other data parse: \code{\link{list.drop}}

Other data parse: \code{\link{list.drop}}
}

