% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{packs.hush}
\alias{packs.hush}
\title{Silent iterator for loading packages and dependencies.}
\usage{
packs.hush(pack_list = NULL, iters = "library")
}
\description{
\code{packs.hush}
}
\seealso{
packs.launch

Other startup functions: \code{\link{adf}},
  \code{\link{chain.df}}, \code{\link{dev.opts_in}},
  \code{\link{parse.file_ext}}, \code{\link{ul.as}}
}

