% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{parse.file_ext}
\alias{parse.file_ext}
\title{Parse a file extension from a relative path}
\usage{
parse.file_ext(files = NULL, with_paths = FALSE)
}
\arguments{
\item{files}{takes a string or list of file names.}

\item{with_paths}{logical Do the file names contain paths.
Relatively useless but kept for backwards compatibility.}
}
\value{
a DataFrame with two columns files and ext for extensions
}
\description{
\code{parse.file_ext}
}
\details{
Given a file name, path or relative path this will extract the
file extensions and parse properly into a data frame object that can
then be used as a ref table.
}
\examples{
parse.file_ext(c('this.rds','this.min.css','this.html'))
> @format
> A data frame with 3 observations on the following 2 variables.
> \\describe{
>   \\item{files}{a character vector}\\emph{ex:this.rds}
>   \\item{ext}{a character vector}\\emph{ex:.rds}
> }
}
\seealso{
Other startup functions: \code{\link{adf}},
  \code{\link{chain.df}}, \code{\link{dev.opts_in}},
  \code{\link{packs.hush}}, \code{\link{ul.as}}
}

