% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{seq.nth}
\alias{seq.nth}
\title{Generate numeric cuts for parsing data or text constraints}
\usage{
\method{seq}{nth}(x, nth)
}
\description{
\code{seq.nth}
}
\examples{

lapply(1:5,function(i)seq.nth(i,nth = 25))


}

