% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_validations.R
\name{swap.is_invalid}
\alias{swap.is_invalid}
\title{\code{swap.is_invalid}}
\usage{
swap.is_invalid(xx, new_val = NULL, invalid_type = c("null", "na", "false",
  "true", "empty"))
}
\description{
\code{swap.is_invalid}
}

