% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_utils.R
\name{ul.as}
\alias{ul.as}
\title{Unlist to alt format}
\usage{
ul.as(x, new_class)
}
\arguments{
\item{new_class}{string The class of the returned object
The options are}
}
\description{
\code{ul.as}
}
\details{
A quick way to convert object classes in a chain command.
}
\seealso{
Other startup functions: \code{\link{adf}},
  \code{\link{chain.df}}, \code{\link{dev.opts_in}},
  \code{\link{packs.hush}}, \code{\link{parse.file_ext}}
}

